refactor: dividing user dtos in folders
