+++ api/src/auth/dto/auth-user/auth-user.dto.ts
@@ -1,5 +1,5 @@
 import { IntersectionType, PickType } from "@nestjs/swagger";
-import { ApplicationUserDto } from "src/users/dto/application-user.dto";
+import { ApplicationUserDto } from "src/users/dto/user/application-user.dto";
 
 export class BaseAuthUserDto extends PickType(ApplicationUserDto, [
   "id",
