+++ api/src/courses/dto/course-owner.dto.ts
@@ -1,5 +1,5 @@
 import { PickType } from "@nestjs/swagger";
-import { ApplicationUserDto } from "src/users/dto/application-user.dto";
+import { ApplicationUserDto } from "src/users/dto/user/application-user.dto";
 
 export class CourseOwnerDto extends PickType(ApplicationUserDto, [
   "id",
