+++ api/src/users/dto/physiotherapist/physiotherapist-profile.dto.ts
@@ -1,6 +1,6 @@
 import { IntersectionType, PickType } from "@nestjs/swagger";
 import { PhysiotherapistDto } from "./physiotherapist.dto";
-import { ApplicationUserDto } from "./application-user.dto";
+import { ApplicationUserDto } from "../user/application-user.dto";
 
 export class PhysiotherapistProfileDto extends IntersectionType(
   PhysiotherapistDto,
